// src/types.ts
export type ResourceID = string;
export type CategoryID = string;
export type EncounterID = string;
export type EnemyID = string;
export type AllyID = string;
export type ActionID = string;
export type ItemID = string;
export type SkillID = string;
export type PartyUnitID = string;
export type EnemyUnitID = number;
export type ProgressionID = string;

export const MainCharacterPartyUnitID = "player";
export type ThemeID = 'light' | 'dark' | 'system';
export type Tab = 'actions' | 'encounter' | 'inventory' | 'config';
export const DefaultTab = 'actions';

export interface CategoryConfig {
    id: CategoryID;
    name: string;
    description: string;
}

export interface Prerequisite {
    // Resource prerequisites
    resourceID?: ResourceID;
    minValue?: number;
    maxValue?: number;
    minMax?: number;
    maxMax?: number;
    valueGreaterThanZero?: boolean;
    valueLessThanMax?: boolean;
    valueEqualToMax?: boolean;
    // Action prerequisites
    actionID?: ActionID;
    actionMinExecutions?: number;
    actionMaxExecutions?: number;
    // Unit prerequisites
    unit?: 'mainCharacter' | 'enemy' | 'enemies' | 'ally' | 'allies' | 'all';
    unitStat?: 'health' | 'mana';
    // Progression prerequisites
    progressionID?: ProgressionID;
    metaProgressionID?: ProgressionID;
}